Move deprecated API into compat
