+++ src/base/colors/i-color-helper.ts
@@ -3,4 +3,5 @@ import { BaseAccessor } from '../../core/types';
 export interface IColorHelper {
     getColor(d, i?: number): string;
     colorAccessor: BaseAccessor<string>;
+    share(colorAccessor: BaseAccessor<string>): IColorHelper;
 }
